Update patch set 1
Patch Set 1:

We can't have a hal_x_client attribute because, in general, different users of the same HAL may need different subsets of rules. For example, let's say a HAL offers a more privileged operation and a less privileged operation and some clients are not supposed to use the more privileged operation.

Patch-set: 1