Update patch set 1
Patch Set 1:

(2 comments)

The line:

  allow init { dev_type -kmem_device -port_device }:chr_file { read open setattr };

feels premature to remove, since there was no auditallow in place for it. As a result, you can't add the neverallow rule either, yet.  :-(

Everything else should be fine, and we probably should get some auditallows in place for that dev_type rule, to see if it's actually used.

Patch-set: 1