Update patch set 3
Patch Set 3: Code-Review+2

When you run the binary from adb shell, it runs in the "su" or "shell" context. However, when you run the update_verifier from init, it runs in the "update_verifier" context. That explains the different denials. SELinux rules are based not only on what executable is executed, but who is executing the executable.

Patch-set: 3
Label: Code-Review=+2