Use _client and _server for Audio HAL policy
This starts the switch for HAL policy to the approach where:
* domains which are clients of Foo HAL are associated with
  hal_foo_client attribute,
* domains which offer the Foo HAL service over HwBinder are
  associated with hal_foo_server attribute,
* policy needed by clients of Foo HAL only when the HAL runs
  in so-called passthrough mode (in clients' processes rather than
  outside) is conditionally included, at build time.
  NOTE: such policy is currently always included while we're working
  on an acceptable way to automatically detect at build time whether
  the policy needs to be included.

A benefit of associating all domains which are clients of Foo HAL with
hal_foo_client, is that this removes the need for device-specific
policy to be able to reference these domains directly (in order to
add device-specific allow rules). Instead, device-specific policy only
needs to reference hal_foo_client and should no longer need to care
which particular domains on the device are clients of Foo HAL. This
can be seen in simplification of the rules for audioserver domain which
is a client of Audio HAL whose policy is being restructured in this
commit.

This commit uses Audio HAL as an example to illustrate the approach.
Once this commit lands, other HALs will also be switched to this
approach.

The approach for creating conditional policy leads to an unfortunate
duplication of rules. However, it was perceived as being simpler to
understand and easier to analyze in the long run. Moreover, going
forward, the need for such conditional policy will be reduced if not
completely removed.

Test: in progress -- this is just to demonstrate the appproach.
Bug: 34170079
Change-Id: I2597a046753edef06123f0476c2ee6889fc17f20