Update patch set 5
Patch Set 5:

> hal_foo_client is any SELinux domain which is a client of the Foo
 > HAL. For example, in the case of Audio HAL, the only client is
 > audioserver domain. The two main reasons we want hal_foo_client
 > are: (1) this lets us clean up rules of client domains which are
 > needed there only when the HAL runs in passthrough mode, and (2)
 > this lets us hide most system domains (e.g, audioserver) from
 > vendor/device-specific policy because those domains were previously
 > needed only because vendor image / device-specific libraries
 > injected code/behavior into those domains and thus forced the need
 > to have additional allow rules, which can now be written against
 > hal_foo_client instead.

I see. Is it then correct that 'hal_audio_client' and 'audioserver' (for the case of audioserver) are equivalent, and the distinction is just a logical one -- that is, if a certain permission is needed to support hal operation then it's attributed to 'hal_audio_client', whereas if it's inherent to the server itself, it's attributed to 'audioserver'.

Then I'm wondering what will be displayed in SELinux denial messages in case when neither permission nor denial rule can be found.

Patch-set: 5