Update patch set 2
Patch Set 2:

> Why don't we just dontaudit the denial, instead of allowing it?
 > Does this actually enable any functionality we didn't have before?
 > Other than suppressing the denial, I don't see what this change
 > enables... I fear we're making a change without understanding *WHY*
 > we're making the change.

Like I said earlier, I know why we're making the change; it's because we're accessing hwservicemanager to retrieve references to HIDL services. I just looked at the denial again, and in particular it looks like this is trying to get the graphics mapper HIDL service; since that is in passthrough mode, we make a call to hwservicemanager to let it know that we've obtained a reference to it.

If your question is why this needs access to HIDL services, I haven't looked into it, but since it's a bootanimation trying to render and produce sound, it doesn't surprise me that it does.

Patch-set: 2
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=0