Update patch set 2
Patch Set 2:

> sorry. Thanks for clarifying....

To clarify, if bootanim only asks for the mapper service, which is always passthrough, we won't need any additional binder_call() here. The call to hwservicemanager is just to let it know we've loaded the mapper in passthrough mode, so this can be shown as a client of the mapper service in the 'lshal' output.

Patch-set: 2
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=0