Update patch set 2
Patch Set 2:

For the record, the reason the policy was written the way it was previously was to ensure we were never calling restorecon on the same set of files twice. The goal was to ensure that a label could only go one way - from the generic "debugfs" or "sysfs" type to a more specific type. 

It should be sufficient to resolve this bug to only grant getattr, and not grant relabelfrom.

Patch-set: 2
CC: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>