Update patch set 1
Patch Set 1:

The intention is to ensure that this file is visible to apps running API level 25 or lower (Nougat). untrusted_app_25.te is the location to put the allow rule in that case. By omitting the rule from untrusted_app_all, we are removing this rule for API level 26 and above. So this will restrict only API level 26 and above, and have no effect on API level 25.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0