Update patch set 1
Patch Set 1:

I'm curious why 'allow modprobe self:capability sys_module' is not required in recovery mode? Was it verified that modules are actually loading successfully? The denial for this is suppressed so if you're just going by the logs you will not see it.

Patch-set: 1
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>