Update patch set 4
Patch Set 4:

Let's stop churning on this CL and get the core of it in. Please remove the new label on /proc/cmdline, leave the new label on /proc/modules. I will +2 that change, and we can continue to discuss other changes.

Patch-set: 4
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>