Update patch set 2
Patch Set 2:

This patch is just a representative of what we have to do to grant rawio capability to the service and build. Unless CTS is rebuilt with this patch (which it wont be), then CTS still fails. This wouldn't actually ever need to be merged, and this patch doesn't really help the reviewer of the waiver. The waiver is we need cap sys rawio and this violates the CTS neverallow checker because OEMs cannot grant this permission to our services.

If you wanted a generic way to do this, which maybe could be merged would be to add:

1. Declare a new attribute sys_rawio_domain
2. Modify the neverallow:
    -sysrawdomain instead of our service.
3. Then oem's can just attribute services that need this capability..

The amount of services that have needed this have been very small, so perhaps like world accessible /dev devices, Google just wants for folks to file an exception so they get a chance to review the use case as well.

Patch-set: 2
Reviewer: Gerrit User 1013433 <1013433@85c56323-6fa9-3386-8a01-6480fb634889>