Update patch set 4
Patch Set 4:

That is because the platform libraries (libvulkan, libui, ...), which are loaded in app processes, are calling configstore HAL.

I think having system_server in between apps and configstore HAL is overkill. The system_server will do nothing but blindly passing calls. 

Configstore HAL is readonly and is designed to be a (future) replacement of the sysprop. So, I think app should be able to talk to it directly.

Patch-set: 4
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0