Update patch set 3
Patch Set 3:

As I understand it update_engine in recovery still statically links the legacy HAL implementation (as opposed to calling getService() to access the new implementation in passthrough mode). Will that be enough for it to be granted hal_bootctl permissions after this change?

Patch-set: 3
Reviewer: Gerrit User 1115183 <1115183@85c56323-6fa9-3386-8a01-6480fb634889>