Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> nfc is an app, and is potentially stored on the vendor partition. We should consider leaving this in public.

It's currently defined strictly as a system image app -- see private/seapp_contexts. Thus, its policy should definitely be in private/nfc.te. If somebody for some reason decides to have the app on vendor image, they'd need to add an entry to private/seapp_contexts and the appropriate rules for that APK which will be stored on the vendor partition. Thus, its policy should then be in vendor/nfc.te. Arguable, that app should not be offering the nfc_service Binder service in that case, because it won't be updated in framework-only OTA  updates...

Patch-set: 2
Reviewer: Gerrit User 1018773 <1018773@85c56323-6fa9-3386-8a01-6480fb634889>