Update patch set 18
Patch Set 18: Code-Review-1

Remains as an RFC to the possibility of an logcat domain.

Likely will cause problems whereever liblogcat.so or the reader interface for liblog.so is used.

The access to the shell data files should be removed ...

Patch-set: 18
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1