Update patch set 1
Patch Set 1:

I agree that this is probably too permissive -- but I also don't have a handle on all the permissions that these apps actually use, nor do I think anyone really does -- the apps do many different things and some of the code is pretty old. If I started out with a blank domain, I would most likely end up removing a necessary permission somewhere and causing problems later on (which is what happened with the linked bug) and then we'd have to revert the domain change again. I was hoping since this is a strictly less permissive change that we'd be able to merge it first, perhaps with auditallows on candidates for removal, and harden them over time.

Patch-set: 1
Reviewer: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>