Update patch set 1
Patch Set 1:

Why revert? Test your change as extensively as you can, run CTS on it, and then address the remaining bugs as they come in. I don't know a better way of doing this, but the current patchset is not acceptable. Granting mediaprovider permission to write OTAs and talk to update_engine and recovery is clearly wrong. And those are just the ones I'm familiar with.

Patch-set: 1
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>