Update patch set 1
Patch Set 1:

Hi Satish,

I understand why you want to minimize device-specific code, and I appreciate that.  Hopefully I can explain the background a little better.

The idea of Treble is to make a clean separation between the Android system image and the underlying implementation.  This should make testing and verification much simpler, and reduce dependencies and breakages.

The default implementation in hardware/interfaces/bluetooth/ is just a wrapper around the old (deprecated) interface.  The new interface just sends and receives HCI packets.  As long as the platform can handle HCI packets, the stack should be able to work with that.

VtsHalBluetoothV1_0Target is the first compatibility test.  It opens and closes the interface, sends an HCI reset, performs loopback tests, ...  It should be helpful for debugging platform-specific code.

Every time we change the default implementation, we risk breaking other platforms.  At this point, they are all working, so the default implementation is effectively frozen.

I agree that this means there will be more duplication of platform-specific code.  That's unfortunate, but I think it is outweighed by the guarantee that no one else can break your platform if your HAL is compliant.

Thanks,
Myles

Patch-set: 1
Reviewer: Gerrit User 1076123 <1076123@85c56323-6fa9-3386-8a01-6480fb634889>