Rework neverallow for /data execute permission
When cherrypicking and merging the change in https://android-review.googlesource.com/#/343713/,
based on the request on https://android-review.googlesource.com/#/342815/,
the partial patch backport request from Android 7.1 to Android 7.0 became a full patch merge.

Please let us know if the partial patch backporting of Android 7.1 to Android 7.0 could be re-evaluated,
and the change in this commit could be checked.

Change-Id: I1c7a99b29ecbeec70a0dc69cd25e029f53aab29f