Update patch set 1
Patch Set 1:

> We're running address sanitizer on the host now? Nice!

It's been on for SANITIZE_HOST builds for a while, but we had to take it down from the main go/ab pages because it can be disruptive for buildcops, etc., since most people don't consider leaks or non-breaking violations to be critical to fix. As these things get fixed now, though, we can work on turning this back on full-time.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>