Update patch set 1
Patch Set 1:

I'm afraid due to the current branching situation and the new SELinux policy API and versioning, we can't accept this until we've made sure the backwards-compatibility set-up is there.  Currently, for instance, if we accepted this into the platform policy and it compiled with non-platform policy that was unaware of this permission, a process which requires it could be denied functionality that it had previously used.

That being said, the likelihood of such a device running a 4.12 kernel is very low, but I think we should follow the same process regardless.

Patch-set: 1
Reviewer: Gerrit User 1026058 <1026058@85c56323-6fa9-3386-8a01-6480fb634889>