Update patch set 1
Patch Set 1:

I'm not sure how well I can clarify this without talking about branch specifics, and in this case I agree that it is rather academic (since this kernel is not yet in the wild).  

I will say that in O we have split policy between the /system and /vendor partitions, where it is expected that it be possible for /system to be updated without an update to /vendor.  Thus, in a general sense, if a permission or class is added in version O+ to the policy present on /system, none of the policy that was written as part of /vendor will be able to add the appropriate allow rules to enable use of the new permission or class, which in the general case could result in previously allowed behaviors being denied.  We need to avoid this.  

A way to allow this, and the one I chose to use, is to use a new mapping file targeting O to add policy indicating that these new permissions/classes are basically "allow all," so that when an O+ system image with a new access_vectors entry is used, the components relying on the /vendor O policy will continue to function.  This file will be dropped once an O+ /system is being used with an O+ /vendor image.

We are currently in such a situation, where this change would not be reflected in policy written for O, but we also have not yet established our compatibility file between O+ (master in this case) and O, to allow for the addition of such rules.  Thus, even though we know that no functionality will be broken by this, since the permission will never be checked due to the missing component in the kernel, I would still like to go through the proper procedure.  Thus, I would plan on accepting this once our O policy is frozen and the mapping can be generated.

Patch-set: 1
Reviewer: Gerrit User 1026058 <1026058@85c56323-6fa9-3386-8a01-6480fb634889>