Update patch set 1
Patch Set 1:

This is not in device policy on any device.

The reason being that it's already in core policy, and any device that has this both in device policy and common policy doesn't boot because the two policies cannot coexist (as described in the commit message).

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>