Update patch set 4
Patch Set 4:

Just FYI, Fedora avoids compatibility problems with new kernel permission checks by setting handle_unknown to allow when building their policies (via the -U option to checkpolicy or secilc). Then new permissions are implicitly allowed under old policies and only enforced when you update to a policy that defines them. Personally I prefer setting it to deny (the default) from a security POV but you might be driven to switching it to avoid compat issues especially with your split policies.
I don't think the CTS currently checks it, so that might be something you want to add if you want to ensure that it stays with a particular value (allow or deny).
Also, another potentially useful CTS test would be to check for any SELinux:.*not defined and/or "SELinux: the above unknown classes and permissions will be..." messages in dmesg/logcat, to require that the policy always defines all of the classes/permissions supported by the kernel.

Patch-set: 4
Reviewer: Gerrit User 1010111 <1010111@85c56323-6fa9-3386-8a01-6480fb634889>