Update patch set 1
Patch Set 1:

> This change will allow network to work, however there are few more
 > things that SELinux is complaining:

In the future, please include the "avc denied" message, not the output to audit2allow. The denial message includes more information necessary to help debug this.

 > allow netd kernel:system module_request;

netd is issuing a request which is causing the kernel to attempt to load a kernel module. In all the cases I've seen, this is an attempt to load a kernel module which doesn't exist. All modules netd needs should be compiled into the kernel. This is a dontaudit candidate.

 > allow netd proc_net:dir { add_name write };
 > allow netd proc_net:file create;

netd is trying to create a file in /proc/net. The /proc/net filesystem isn't a filesystem where new entries can be created. Even if this was allowed, the underlying filesystem would eventually reject this attempt to create a file. Best solution would be to identify what file is attempting to be created and fix the relevant code. Allowing this in policy isn't appropriate.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>