Update patch set 2
Patch Set 2:

I'd prefer we stop moving rules to private.  The public/private distinction is best for types, not for rules.  Moving a rule to private means that it will not be duped on vendor partitions, and theoretically shouldn't be relied upon by vendors, but without a way to test that, it's frail.

Patch-set: 2
Reviewer: Gerrit User 1026058 <1026058@85c56323-6fa9-3386-8a01-6480fb634889>