Update patch set 2
Patch Set 2: Code-Review+1

Thanks. It's entirely possible that CTS team will ask you to reintroduce it, but I see no reason to keep this change out of MR0. It would be weird to have it in master, and N-MR0 but not N-MR1

Patch-set: 2
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1