Vold no execute_no_trans
For the same reason as init should not be passed to help processes
restrict so they are enforced to have their own domain.

Please note that this will also require changes to the code
(vold and script). For example the toolbox usage of rm and cp need to
be fixed and shell scripts need exec domains. Init has done this with
built in functions.

Bug: 64791922
Test: Manual
Change-Id: Ide550d00dfaab42ab0aa74fec93d6055f1a35aa1