bootstat: lock down *_boot_reason_prop
Add series of neverallow rules to restrict components for reading or
writing bootloader_boot_reason_prop, system_boot_reason_prop and
last_boot_reason_prop to trusted set.

The policy is that bootloader_boot_reason_prop (ro.boot.bootreason)
has a compliance issue due to the sheer momentum of near unparseable
content as filed by the bootloader and is only to be accessible to a
series of responsible system components.  It can be inacurate as it
provides no means to evaluate a shutdown.

last_boot_reason_prop (persist.sys.boot.reason) contains inaccurate
information as it is only valid after a controlled reboot or
shutdown.  Since the information could be false, we do not want to
support it as an open API, so we again block access to only
responsible components.

system_boot_reason_prop (sys.boot.reason) is a canonical boot reason
that takes into account parsing bootloader_boot_reason_prop,
boot_loader_boot_reason_prop and other system and HAL generated hints
to determine a parseable and most accurate reason for the last time
the system was rebooted.

For now the policy for system_boot_reason_prop is to audit users of
the API, and on a need to know basis via selinux rules, but could
very well at some point in the future be opened up to all should
this become a scaling issue.  If vendors need their components to
access the boot reason, they need to comply first with CTS tests with
regards to controlled reboot messaging and in turn read the
system_boot_reason_prop for the canonical information.

The controlled reboot APIs include:
- android_reboot(ANDROID_RB_<TYPE>, int flag, const char* reason)
- PowerManagerService.lowLevelShutdown(String reason);
- PowerManagerService.lowLevelReboot(String reason);
- ShutdownThread.shutdown(context, String reason, boolean confirm);
- ShutdownThread.reboot(context, String reason, boolean confirm);
- PowerManager.shutdown(boolean confirm, String reason, boolean wait);
- PowerManager.reboot(String reason);

Any others (including the direct linux reboot syscall) create problems
for generating an accurate canonical boot or shutdown reason.

Test: compile
Bug: 63736262
Change-Id: I2e5e55bbea1c383c06472eb2989237cfeb852030