Remove execute_no_trans from vold
For the same reason as init should not be passed to helper processes
also restrict this for vold and enforce the helpers to have their own
domain.

To be able to run tools we introduce a new vold_tool domain. It supposed
to be trusted but we can strip down a lot of rights from vold to be able
to perform the tool tasks.

Bug: 64791922
Test: Manual
Change-Id: Ide550d00dfaab42ab0aa74fec93d6055f1a35aa1