Update patch set 4
Patch Set 4:

We did a sanity test and the change seems to work OK on a Nexus 6P running the (as of yesterday) latest AOSP code.

The upload was just a rebase to the latest changes in the git. No actual functional modification was done.

Patch-set: 4
Reviewer: Gerrit User 1037953 <1037953@85c56323-6fa9-3386-8a01-6480fb634889>