Update patch set 1
Patch Set 1:

The issue, I think is that the co-dr branch was moved to be a "release" branch, which then triggers using it to use the given release code (26.0).  This is the same one that is in AOSP master, because  that hasn't been updated yet to a newer version, so having a compat file for 26.0 conflicts with the one that gets built for the system image.

Patch-set: 1
Reviewer: Gerrit User 1026058 <1026058@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0