Update patch set 1
Patch Set 1:

(1 comment)

> Ok, so perhaps I lack context here. Why are we moving this policy
 > to device-specific policy? There is nothing device-specific about
 > this code that I can see. Do you have a bug or a design doc with
 > more context?

I don't think the CL is trying to do that, it is in fact only trying to make sure netd only has permissions that it requires and nothing else. The sysfs_net access in pixel policy mentioned above is probably a bug that must/should be fixed in the policy to make sure rules granting that access is added in platform sepolicy.

This is part of our effort to specifically label interfaces used by the platform in sys/proc and lock access to everything else. So, netd having access to generic 'sysfs' is one of those things. The umbrella bug is at - b/65643247.

Once we are done, what this will give us is the list of "kernel interfaces" that Android depends upon and that is valuable because then we can now monitor them, add tests for them in VTS and ensure tests are added for any new usages.

Patch-set: 1
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>