Update patch set 9
Patch Set 9: Code-Review+1

(1 comment)

(unrelated meta comment) domain.te is too big :(, may be take the neverallows out of it and put it in a separate file.

+1 to make sure others have looked at this too. trong@, can you please make sure both treble / non-treble devices build with this?

There is one concern I have with the overall change, What is the mechanism that we provide to vendors / OEMS to adda a 'coredomain' process that now wants to use a new sysfs file?

or Are we asking them to make a HAL for it ?

Patch-set: 9
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1